# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WIZIA API Documentation",
        "description": "Documentation de l'API WIZIA",
        "contact": {
            "email": "contact@wizia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/stripe/create-payment-intent": {
            "post": {
                "tags": [
                    "Payment"
                ],
                "summary": "Créer un PaymentIntent Stripe et mettre à jour l'abonnement utilisateur",
                "operationId": "3062d21bd25db7d4fb068c2fede2d81b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "IdUser",
                                    "email",
                                    "nom"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant en centimes",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2000
                                    },
                                    "IdUser": {
                                        "description": "ID de l'utilisateur",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "email": {
                                        "description": "Email pour le reçu",
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "nom": {
                                        "description": "Nom de l'abonnement (Free, Premium, Professionnel)",
                                        "type": "string",
                                        "example": "Premium"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "PaymentIntent créé avec succès et abonnement mis à jour"
                    },
                    "400": {
                        "description": "Erreur de validation ou abonnement non valide"
                    },
                    "404": {
                        "description": "Utilisateur ou abonnement introuvable"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
        "/stripe/abonnement/{id}": {
            "get": {
                "tags": [
                    "Payment"
                ],
                "summary": "Récupérer le type d'abonnement d'un utilisateur",
                "operationId": "3c5ee1fd6529e751eca744f6a61f308f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type d'abonnement retourné (isFree, isPremium, isProfessionnel)"
                    },
                    "400": {
                        "description": "Aucun type d’abonnement valide"
                    },
                    "404": {
                        "description": "Utilisateur ou abonnement non trouvé"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            }
        },
...
    },
...
}